+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - ThreadBehaviour to define how a thread is created
 - Allow benchmarks to be run on the MainThread
+- OnExecuted to run a delegate after each task run
  
 ### Changed
 - Added IDisposable to IThreadSessionHandler